+++ index.ios.js
@@ -1,9 +1,12 @@
-import Containers from './src/lib/containers'
-import Components from './src/lib/components'
-import Routes from './src/lib/relay/routes'
+// First things first.
+import "./src/lib/error_reporting"
 
-import './src/lib/relay/config'
-import './src/lib/app_registry'
+import Containers from "./src/lib/containers"
+import Components from "./src/lib/components"
+import Routes from "./src/lib/relay/routes"
+
+import "./src/lib/relay/config"
+import "./src/lib/app_registry"
 
 export default {
   Containers,